# backend/openapi.json
{
    "openapi": "3.1.0",
    "paths": {
        "/auth/sign-up": {
            "post": {
                "operationId": "Auth/signUp",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignUpDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/sign-down": {
            "post": {
                "operationId": "Auth/signDown",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignDownDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Status"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/sign-in": {
            "post": {
                "operationId": "Auth/signIn",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/AppCredentials"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/sign-out": {
            "post": {
                "operationId": "Auth/signOut",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Status"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/refresh-token": {
            "post": {
                "operationId": "Auth/refreshToken",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshTokenDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/AppCredentials"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/verify-email": {
            "post": {
                "operationId": "Auth/verifyEmail",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyEmailDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Status"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/forgot-password": {
            "post": {
                "operationId": "Auth/forgotPassword",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Status"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/confirm-password": {
            "post": {
                "operationId": "Auth/confirmPassword",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfirmPasswordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Status"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/auth/change-password": {
            "post": {
                "operationId": "Auth/changePassword",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Status"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Auth"
                ]
            }
        },
...
    },
    "info": {
        "title": "API",
        "description": "",
        "version": "1.0.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "ApiError": {
                "type": "object",
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "status": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    },
                    "detail": {
                        "type": "object"
                    }
                },
                "required": [
                    "code",
                    "status",
                    "message"
                ]
            },
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "string"
                    },
                    "endpoint": {
                        "type": "string"
                    },
                    "docs": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    },
                    "error": {
                        "$ref": "#/components/schemas/ApiError"
                    }
                },
                "required": [
                    "version",
                    "endpoint",
                    "docs",
                    "timestamp"
                ]
            },
            "AppCredentials": {
                "type": "object",
                "properties": {
                    "identityToken": {
                        "type": "string"
                    },
                    "accessToken": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "identityToken",
                    "accessToken",
                    "refreshToken"
                ]
            },
...
            "ErrorCode": {
                "type": "string",
                "oneOf": [
                    {
                        "title": "GenericError",
                        "const": "generic_error"
                    },
                    {
                        "title": "InvalidRequest",
                        "const": "invalid_request"
                    },
                    {
                        "title": "ValidationError",
                        "const": "validation_error"
                    },
                    {
                        "title": "Unauthorized",
                        "const": "unauthorized"
                    },
                    {
                        "title": "Forbidden",
                        "const": "forbidden"
                    },
                    {
                        "title": "NotFound",
                        "const": "not_found"
                    },
                    {
                        "title": "EmailTaken",
                        "const": "email_taken"
                    },
                    {
                        "title": "ConfigurationError",
                        "const": "configuration_error"
                    }
                ]
            },
            "Role": {
                "type": "string",
                "oneOf": [
                    {
                        "title": "Admin",
                        "const": "admin"
                    }
                ]
            },
            "Status": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "status"
                ]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "birthdate": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "emailVerified": {
                        "type": "boolean"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "phoneVerified": {
                        "type": "boolean"
                    },
                    "locale": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "required": [
                    "id",
                    "name",
                    "surname",
                    "birthdate",
                    "email",
                    "emailVerified",
                    "phone",
                    "phoneVerified",
                    "locale",
                    "timezone",
                    "icon",
                    "roles"
                ]
            },
            "ChangePasswordDto": {
                "type": "object",
                "properties": {
                    "currentPassword": {
                        "type": "string"
                    },
                    "newPassword": {
                        "type": "string"
                    }
                },
                "required": [
                    "currentPassword",
                    "newPassword"
                ]
            },
            "ConfirmPasswordDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "code",
                    "password"
                ]
            },
            "ForgotPasswordDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "email"
                ]
            },
            "RefreshTokenDto": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "refreshToken"
                ]
            },
            "SignDownDto": {
                "type": "object",
                "properties": {
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "password"
                ]
            },
            "SignInDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "SignUpDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
...
            "VerifyEmailDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    }
                },
                "required": [
                    "code"
                ]
            }
        }
    }
}